+++ GFramework.Core/extensions/ContextAwareQueryExtensions.cs
@@ -42,24 +42,6 @@ public static class ContextAwareQueryExtensions
         return context.SendQuery(query);
     }
 
-    /// <summary>
-    ///     [Mediator] 异步发送查询并返回结果
-    /// </summary>
-    /// <typeparam name="TResponse">查询响应类型</typeparam>
-    /// <param name="contextAware">实现 IContextAware 接口的对象</param>
-    /// <param name="query">要发送的查询对象</param>
-    /// <param name="cancellationToken">取消令牌,用于取消操作</param>
-    /// <returns>包含查询结果的ValueTask</returns>
-    /// <exception cref="ArgumentNullException">当 contextAware 或 query 为 null 时抛出</exception>
-    public static ValueTask<TResponse> SendQueryAsync<TResponse>(this IContextAware contextAware,
-        Mediator.IQuery<TResponse> query, CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(contextAware);
-        ArgumentNullException.ThrowIfNull(query);
-
-        var context = contextAware.GetContext();
-        return context.SendQueryAsync(query, cancellationToken);
-    }
 
     /// <summary>
     ///     异步发送一个查询请求
